removed custom settings, shuttle display with arrows
