+++ ShuttleTrackeriOS/Extensions.swift
@@ -1,0 +1,31 @@
+//
+//  Extensions.swift
+//  ShuttleTrackeriOS
+//
+//  Created by Andrew Qu on 2/15/19.
+//  Copyright © 2019 WTG. All rights reserved.
+//
+
+import Foundation
+import MapKit
+
+extension ViewController: MKMapViewDelegate {
+    func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
+        guard annotation is MKPointAnnotation else { return nil }
+        
+        let identifier = "Annotation"
+        var annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: identifier)
+        
+        if annotationView == nil {
+            annotationView = MKPinAnnotationView(annotation: annotation, reuseIdentifier: identifier)
+            annotationView!.canShowCallout = true
+        } else {
+            annotationView!.annotation = annotation
+        }
+        
+        return annotationView
+    }
+}
+
+
+
