+++ ShuttleTrackeriOS/ShuttleDisplay.swift
@@ -1,0 +1,19 @@
+//
+//  ShuttleDisplay.swift
+//  ShuttleTrackeriOS
+//
+//  Created by Andrew Qu on 2/14/19.
+//  Copyright © 2019 WTG. All rights reserved.
+//
+
+
+import Foundation
+import MapKit
+
+
+    
+class ShuttleDisplay {
+    
+    
+    
+}
